# Convert score from byte array to uint64