--- a/score.go
+++ b/score.go
@@ -1,8 +1,7 @@
 package dieci
 
 import (
-	"encoding/binary"
-	"encoding/hex"
+	"strconv"
 
 	"github.com/OneOfOne/xxhash"
 )
@@ -11,16 +10,15 @@ import (
 const scoreSize = 8
 
 // Score is type alias for score representation
-type Score [scoreSize]byte
+// type Score [scoreSize]byte
+type Score uint64
 
 func (s Score) String() string {
-	return hex.EncodeToString(s[:])
+	return strconv.FormatInt(int64(s), 16)
 }
 
 // MakeScore creates a score for a given data block
 func MakeScore(b []byte) Score {
 	sum := xxhash.Checksum64(b)
-	score := Score{}
-	binary.BigEndian.PutUint64(score[:], sum)
-	return score
+	return Score(sum)
 }