Don't confuse representation with serialization 😉 

```
$ benchcmp before.bench after.bench
benchmark                old ns/op     new ns/op     delta
BenchmarkOpenIndex-4     85194920      72041524      -15.44%
BenchmarkOpen-4          61195548      60705285      -0.80%
BenchmarkWrite-4         12458         12381         -0.62%
BenchmarkRead-4          1148          1120          -2.44%

benchmark                old allocs     new allocs     delta
BenchmarkOpenIndex-4     9680           9696           +0.17%
BenchmarkOpen-4          9654           9665           +0.11%
BenchmarkWrite-4         1              1              +0.00%
BenchmarkRead-4          1              1              +0.00%

benchmark                old bytes     new bytes     delta
BenchmarkOpenIndex-4     31402293      31406146      +0.01%
BenchmarkOpen-4          31392038      31396432      +0.01%
BenchmarkWrite-4         1246          1247          +0.08%
BenchmarkRead-4          32            32            +0.00%
```